+++ mordant/src/appleMain/kotlin/com/github/ajalt/mordant/internal/MppInternal.apple.kt
@@ -1,0 +1,6 @@
+package com.github.ajalt.mordant.internal
+
+import com.github.ajalt.mordant.internal.syscalls.SyscallHandler
+import com.github.ajalt.mordant.internal.syscalls.SyscallHandlerNativeApple
+
+internal actual fun getSyscallHandler(): SyscallHandler = SyscallHandlerNativeApple